Move via info hooks to related folders
