--- ui/src/hooks/via/index.ts
@@ -1,2 +1,0 @@
-export * from './useEditViaInfo';
-export * from './useRemoveViaInfo';
